refactor: move workspace drive image preparation
