+++ crates/sandbox-fc/src/factory/mod.rs
@@ -6,16 +6,15 @@ mod invariant;
 mod leak_cleaner;
 
 use std::path::{Path, PathBuf};
-use std::time::{Duration, Instant};
+use std::time::Instant;
 
 use async_trait::async_trait;
 use sandbox::{
     Sandbox, SandboxConfig, SandboxError, SandboxFactory, SandboxInitializationPhase,
-    SandboxInvalidStateContext, WorkspaceDriveSeedImage,
+    SandboxInvalidStateContext,
 };
 use tracing::{info, warn};
 
-use crate::command;
 use crate::config::{FirecrackerConfig, FirecrackerDeviceRateLimits};
 use crate::cow_cleanup::CowCleanupOutcome;
 use crate::factory::cleanup_group::{FactoryCleanupGroup, FactoryCleanupTaskKind};
@@ -30,6 +29,7 @@ use crate::network::{NetnsPoolConfig, NetnsPoolHandle};
 use crate::paths::{FactoryPaths, RuntimePaths, SandboxPaths, SockPaths};
 use crate::prerequisites;
 use crate::sandbox::{FirecrackerSandbox, FirecrackerSandboxInit};
+use crate::workspace_drive_image::prepare_workspace_drive_image;
 
 pub(crate) use invariant::InvariantConfig;
 pub use invariant::{PREWARM_SCRIPT, config_hash};
@@ -468,233 +468,6 @@ fn clean_stale_workspace_dir(id: &str, target_workspace: &Path) -> sandbox::Resu
     clean_stale_create_dir(id, "target workspace", target_workspace)
 }
 
-pub(crate) async fn prepare_workspace_drive_image(
-    path: &Path,
-    config: &sandbox::WorkspaceDriveConfig,
-    mut timing: Option<&mut SandboxCreateTiming>,
-) -> sandbox::Result<()> {
-    if let Some(timing) = timing.as_deref_mut() {
-        timing.mark_workspace_drive_present();
-    }
-
-    if let Some(parent) = path.parent() {
-        tokio::fs::create_dir_all(parent)
-            .await
-            .map_err(|e| SandboxError::Initialization {
-                phase: SandboxInitializationPhase::SandboxAllocation,
-                message: format!("create workspace image dir: {e}"),
-            })?;
-    }
-
-    match config.seed_image.as_ref() {
-        Some(WorkspaceDriveSeedImage::Copy(source_image)) => {
-            if let Some(timing) = timing.as_deref_mut() {
-                timing.mark_workspace_seed_image_used();
-            }
-            return copy_workspace_drive_seed_image(
-                path,
-                source_image,
-                workspace_drive_size_bytes(config.size_mb),
-                timing,
-            )
-            .await;
-        }
-        Some(WorkspaceDriveSeedImage::Move(source_image)) => {
-            if let Some(timing) = timing.as_deref_mut() {
-                timing.mark_workspace_seed_image_used();
-            }
-            return move_workspace_drive_seed_image(
-                path,
-                source_image,
-                workspace_drive_size_bytes(config.size_mb),
-            )
-            .await;
-        }
-        None => {}
-    }
-
-    let file = tokio::fs::File::create(path)
-        .await
-        .map_err(|e| SandboxError::Initialization {
-            phase: SandboxInitializationPhase::SandboxAllocation,
-            message: format!("create workspace image {}: {e}", path.display()),
-        })?;
-    file.set_len(workspace_drive_size_bytes(config.size_mb))
-        .await
-        .map_err(|e| SandboxError::Initialization {
-            phase: SandboxInitializationPhase::SandboxAllocation,
-            message: format!("set workspace image size {}: {e}", path.display()),
-        })?;
-    drop(file);
-
-    let path_str = path.to_str().ok_or_else(|| SandboxError::Initialization {
-        phase: SandboxInitializationPhase::SandboxAllocation,
-        message: format!("workspace image path is not UTF-8: {}", path.display()),
-    })?;
-    let stage_started = Instant::now();
-    let result = command::exec_with_timeout(
-        "mkfs.ext4",
-        &["-F", "-q", path_str],
-        Duration::from_secs(60),
-    )
-    .await
-    .map_err(|e| SandboxError::Initialization {
-        phase: SandboxInitializationPhase::SandboxAllocation,
-        message: format!("format workspace image: {e}"),
-    });
-    if let Some(timing) = timing {
-        timing.record_stage_result(
-            SandboxCreateStage::WorkspaceFreshFormat,
-            stage_started,
-            result,
-        )
-    } else {
-        result
-    }?;
-    Ok(())
-}
-
-async fn copy_workspace_drive_seed_image(
-    target: &Path,
-    source: &Path,
-    expected_size_bytes: u64,
-    timing: Option<&mut SandboxCreateTiming>,
-) -> sandbox::Result<()> {
-    validate_workspace_seed_source(source, expected_size_bytes).await?;
-
-    let source_str = source
-        .to_str()
-        .ok_or_else(|| SandboxError::Initialization {
-            phase: SandboxInitializationPhase::SandboxAllocation,
-            message: format!(
-                "workspace seed image path is not UTF-8: {}",
-                source.display()
-            ),
-        })?;
-    let target_str = target
-        .to_str()
-        .ok_or_else(|| SandboxError::Initialization {
-            phase: SandboxInitializationPhase::SandboxAllocation,
-            message: format!("workspace image path is not UTF-8: {}", target.display()),
-        })?;
-
-    let stage_started = Instant::now();
-    let result = command::exec_with_timeout(
-        "cp",
-        &["--sparse=always", "--", source_str, target_str],
-        Duration::from_secs(300),
-    )
-    .await
-    .map_err(|e| SandboxError::Initialization {
-        phase: SandboxInitializationPhase::SandboxAllocation,
-        message: format!("copy workspace seed image: {e}"),
-    });
-    if let Some(timing) = timing {
-        timing.record_stage_result(
-            SandboxCreateStage::WorkspaceSeedSparseCopy,
-            stage_started,
-            result,
-        )
-    } else {
-        result
-    }?;
-
-    validate_workspace_seed_target(target, expected_size_bytes, "copied").await
-}
-
-async fn move_workspace_drive_seed_image(
-    target: &Path,
-    source: &Path,
-    expected_size_bytes: u64,
-) -> sandbox::Result<()> {
-    validate_workspace_seed_source(source, expected_size_bytes).await?;
-    if let Err(e) = std::fs::rename(source, target) {
-        return Err(SandboxError::Initialization {
-            phase: SandboxInitializationPhase::SandboxAllocation,
-            message: format!(
-                "move workspace seed image {} to {}: {e}",
-                source.display(),
-                target.display()
-            ),
-        });
-    }
-    validate_workspace_seed_target(target, expected_size_bytes, "moved").await
-}
-
-async fn validate_workspace_seed_source(
-    source: &Path,
-    expected_size_bytes: u64,
-) -> sandbox::Result<()> {
-    let source_metadata =
-        tokio::fs::symlink_metadata(source)
-            .await
-            .map_err(|e| SandboxError::Initialization {
-                phase: SandboxInitializationPhase::SandboxAllocation,
-                message: format!("read workspace seed image {}: {e}", source.display()),
-            })?;
-    if !source_metadata.is_file() {
-        return Err(SandboxError::Initialization {
-            phase: SandboxInitializationPhase::SandboxAllocation,
-            message: format!(
-                "workspace seed image is not a regular file: {}",
-                source.display()
-            ),
-        });
-    }
-    if source_metadata.len() != expected_size_bytes {
-        return Err(SandboxError::Initialization {
-            phase: SandboxInitializationPhase::SandboxAllocation,
-            message: format!(
-                "workspace seed image size mismatch for {}: expected {} bytes, got {} bytes",
-                source.display(),
-                expected_size_bytes,
-                source_metadata.len()
-            ),
-        });
-    }
-    Ok(())
-}
-
-async fn validate_workspace_seed_target(
-    target: &Path,
-    expected_size_bytes: u64,
-    action: &str,
-) -> sandbox::Result<()> {
-    let target_metadata =
-        tokio::fs::symlink_metadata(target)
-            .await
-            .map_err(|e| SandboxError::Initialization {
-                phase: SandboxInitializationPhase::SandboxAllocation,
-                message: format!("read {action} workspace image {}: {e}", target.display()),
-            })?;
-    if !target_metadata.is_file() {
-        return Err(SandboxError::Initialization {
-            phase: SandboxInitializationPhase::SandboxAllocation,
-            message: format!(
-                "{action} workspace image is not a regular file: {}",
-                target.display()
-            ),
-        });
-    }
-    if target_metadata.len() != expected_size_bytes {
-        return Err(SandboxError::Initialization {
-            phase: SandboxInitializationPhase::SandboxAllocation,
-            message: format!(
-                "{action} workspace image size mismatch for {}: expected {} bytes, got {} bytes",
-                target.display(),
-                expected_size_bytes,
-                target_metadata.len()
-            ),
-        });
-    }
-
-    Ok(())
-}
-
-fn workspace_drive_size_bytes(size_mb: u32) -> u64 {
-    u64::from(size_mb) * 1024 * 1024
-}
-
 fn clean_stale_sock_dir(id: &str, sock_dir: &Path) -> sandbox::Result<()> {
     clean_stale_create_dir(id, "sock dir", sock_dir)
 }
@@ -792,266 +565,8 @@ mod tests {
     use crate::factory::leak_cleaner::LeakCleaner;
     use crate::leaked_resources::LeakedResources;
     use crate::network::{NetnsLease, NetnsPool};
-    use std::io::SeekFrom;
     use std::path::PathBuf;
     use std::sync::{Arc, Mutex};
-    use tokio::io::{AsyncReadExt, AsyncSeekExt, AsyncWriteExt};
-
-    #[test]
-    fn workspace_drive_size_bytes_converts_mib_to_bytes() {
-        assert_eq!(workspace_drive_size_bytes(16), 16 * 1024 * 1024);
-    }
-
-    #[tokio::test]
-    async fn prepare_workspace_drive_image_without_seed_formats_fresh_image() {
-        let tmp = tempfile::tempdir().unwrap();
-        let target = tmp.path().join("nested").join("workspace.ext4");
-        let mut timing = SandboxCreateTiming::new("sandbox".into(), "vm0/default".into());
-
-        prepare_workspace_drive_image(
-            &target,
-            &sandbox::WorkspaceDriveConfig {
-                size_mb: 16,
-                seed_image: None,
-            },
-            Some(&mut timing),
-        )
-        .await
-        .unwrap();
-
-        let metadata = tokio::fs::metadata(&target).await.unwrap();
-        assert_eq!(metadata.len(), workspace_drive_size_bytes(16));
-        assert!(
-            timing
-                .stage_duration_for_test(SandboxCreateStage::WorkspaceFreshFormat)
-                .is_some()
-        );
-        assert!(
-            timing
-                .stage_duration_for_test(SandboxCreateStage::WorkspaceSeedSparseCopy)
-                .is_none()
-        );
-    }
-
-    #[tokio::test]
-    async fn prepare_workspace_drive_image_sparse_copies_seed_image() {
-        let tmp = tempfile::tempdir().unwrap();
-        let source = tmp.path().join("seed.ext4");
-        let target = tmp.path().join("nested").join("workspace.ext4");
-        let marker_offset = 4096;
-        let marker = b"vm0";
-        let mut timing = SandboxCreateTiming::new("sandbox".into(), "vm0/default".into());
-
-        let mut source_file = tokio::fs::File::create(&source).await.unwrap();
-        source_file
-            .set_len(workspace_drive_size_bytes(1))
-            .await
-            .unwrap();
-        source_file
-            .seek(SeekFrom::Start(marker_offset))
-            .await
-            .unwrap();
-        source_file.write_all(marker).await.unwrap();
-        source_file.flush().await.unwrap();
-        drop(source_file);
-
-        prepare_workspace_drive_image(
-            &target,
-            &sandbox::WorkspaceDriveConfig {
-                size_mb: 1,
-                seed_image: Some(WorkspaceDriveSeedImage::Copy(source.clone())),
-            },
-            Some(&mut timing),
-        )
-        .await
-        .unwrap();
-
-        let metadata = tokio::fs::metadata(&target).await.unwrap();
-        assert_eq!(metadata.len(), workspace_drive_size_bytes(1));
-        assert!(
-            timing
-                .stage_duration_for_test(SandboxCreateStage::WorkspaceSeedSparseCopy)
-                .is_some()
-        );
-        assert!(
-            timing
-                .stage_duration_for_test(SandboxCreateStage::WorkspaceFreshFormat)
-                .is_none()
-        );
-
-        let mut target_file = tokio::fs::File::open(&target).await.unwrap();
-        target_file
-            .seek(SeekFrom::Start(marker_offset))
-            .await
-            .unwrap();
-        let mut copied_marker = [0; 3];
-        target_file.read_exact(&mut copied_marker).await.unwrap();
-        assert_eq!(&copied_marker, marker);
-    }
-
-    #[tokio::test]
-    async fn prepare_workspace_drive_image_rejects_seed_image_size_mismatch() {
-        let tmp = tempfile::tempdir().unwrap();
-        let source = tmp.path().join("seed.ext4");
-        let target = tmp.path().join("workspace.ext4");
-
-        let source_file = tokio::fs::File::create(&source).await.unwrap();
-        source_file
-            .set_len(workspace_drive_size_bytes(1) - 1)
-            .await
-            .unwrap();
-        drop(source_file);
-
-        let err = prepare_workspace_drive_image(
-            &target,
-            &sandbox::WorkspaceDriveConfig {
-                size_mb: 1,
-                seed_image: Some(WorkspaceDriveSeedImage::Copy(source)),
-            },
-            None,
-        )
-        .await
-        .unwrap_err();
-
-        match err {
-            SandboxError::Initialization { phase, message } => {
-                assert_eq!(phase, SandboxInitializationPhase::SandboxAllocation);
-                assert!(message.contains("workspace seed image size mismatch"));
-            }
-            other => panic!("expected workspace seed initialization error, got {other:?}"),
-        }
-        assert!(!target.exists());
-    }
-
-    #[tokio::test]
-    async fn prepare_workspace_drive_image_moves_seed_image() {
-        let tmp = tempfile::tempdir().unwrap();
-        let source = tmp.path().join("seed.ext4");
-        let target = tmp.path().join("nested").join("workspace.ext4");
-        let marker_offset = 4096;
-        let marker = b"vm0";
-        let mut timing = SandboxCreateTiming::new("sandbox".into(), "vm0/default".into());
-
-        let mut source_file = tokio::fs::File::create(&source).await.unwrap();
-        source_file
-            .set_len(workspace_drive_size_bytes(1))
-            .await
-            .unwrap();
-        source_file
-            .seek(SeekFrom::Start(marker_offset))
-            .await
-            .unwrap();
-        source_file.write_all(marker).await.unwrap();
-        source_file.flush().await.unwrap();
-        drop(source_file);
-
-        prepare_workspace_drive_image(
-            &target,
-            &sandbox::WorkspaceDriveConfig {
-                size_mb: 1,
-                seed_image: Some(WorkspaceDriveSeedImage::Move(source.clone())),
-            },
-            Some(&mut timing),
-        )
-        .await
-        .unwrap();
-
-        assert!(!source.exists());
-        let metadata = tokio::fs::metadata(&target).await.unwrap();
-        assert_eq!(metadata.len(), workspace_drive_size_bytes(1));
-        assert!(
-            timing
-                .stage_duration_for_test(SandboxCreateStage::WorkspaceSeedSparseCopy)
-                .is_none()
-        );
-        assert!(
-            timing
-                .stage_duration_for_test(SandboxCreateStage::WorkspaceFreshFormat)
-                .is_none()
-        );
-
-        let mut target_file = tokio::fs::File::open(&target).await.unwrap();
-        target_file
-            .seek(SeekFrom::Start(marker_offset))
-            .await
-            .unwrap();
-        let mut moved_marker = [0; 3];
-        target_file.read_exact(&mut moved_marker).await.unwrap();
-        assert_eq!(&moved_marker, marker);
-    }
-
-    #[tokio::test]
-    async fn prepare_workspace_drive_image_rejects_move_seed_image_size_mismatch_without_moving() {
-        let tmp = tempfile::tempdir().unwrap();
-        let source = tmp.path().join("seed.ext4");
-        let target = tmp.path().join("workspace.ext4");
-
-        let source_file = tokio::fs::File::create(&source).await.unwrap();
-        source_file
-            .set_len(workspace_drive_size_bytes(1) - 1)
-            .await
-            .unwrap();
-        drop(source_file);
-
-        let err = prepare_workspace_drive_image(
-            &target,
-            &sandbox::WorkspaceDriveConfig {
-                size_mb: 1,
-                seed_image: Some(WorkspaceDriveSeedImage::Move(source.clone())),
-            },
-            None,
-        )
-        .await
-        .unwrap_err();
-
-        match err {
-            SandboxError::Initialization { phase, message } => {
-                assert_eq!(phase, SandboxInitializationPhase::SandboxAllocation);
-                assert!(message.contains("workspace seed image size mismatch"));
-            }
-            other => panic!("expected workspace seed initialization error, got {other:?}"),
-        }
-        assert!(source.exists());
-        assert!(!target.exists());
-    }
-
-    #[tokio::test]
-    async fn prepare_workspace_drive_image_rejects_move_seed_symlink_without_moving() {
-        let tmp = tempfile::tempdir().unwrap();
-        let real_source = tmp.path().join("real-seed.ext4");
-        let source_link = tmp.path().join("seed-link.ext4");
-        let target = tmp.path().join("workspace.ext4");
-
-        let source_file = tokio::fs::File::create(&real_source).await.unwrap();
-        source_file
-            .set_len(workspace_drive_size_bytes(1))
-            .await
-            .unwrap();
-        drop(source_file);
-        std::os::unix::fs::symlink(&real_source, &source_link).unwrap();
-
-        let err = prepare_workspace_drive_image(
-            &target,
-            &sandbox::WorkspaceDriveConfig {
-                size_mb: 1,
-                seed_image: Some(WorkspaceDriveSeedImage::Move(source_link.clone())),
-            },
-            None,
-        )
-        .await
-        .unwrap_err();
-
-        match err {
-            SandboxError::Initialization { phase, message } => {
-                assert_eq!(phase, SandboxInitializationPhase::SandboxAllocation);
-                assert!(message.contains("workspace seed image is not a regular file"));
-            }
-            other => panic!("expected workspace seed initialization error, got {other:?}"),
-        }
-        assert!(real_source.exists());
-        assert!(source_link.exists());
-        assert!(!target.exists());
-    }
 
     #[tokio::test]
     async fn shutdown_cleans_owned_netns_pool_with_extra_arc_refs() {
